# docs/contracts/rustshare-api-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RustShare REST API",
    "description": "Auto-generated OpenAPI specification for the RustShare file-sharing and sync platform.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
  "paths": {
    "/api/v1/admin/integrations/chat/webhooks": {
      "get": {
        "tags": [
          "Chat Integration"
        ],
        "summary": "GET /api/v1/admin/integrations/chat/webhooks",
        "description": "List registered chat webhook URLs.\nRequires admin privileges.",
        "operationId": "list_chat_webhooks",
        "responses": {
          "200": {
            "description": "Registered chat webhooks",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Chat Integration"
        ],
        "summary": "POST /api/v1/admin/integrations/chat/webhooks",
        "description": "Register a new chat webhook URL for event dispatch.\nRequires admin privileges.",
        "operationId": "register_chat_webhook",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterWebhookRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Webhook registered"
          },
          "400": {
            "description": "Invalid URL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/integrations/chat/events": {
      "post": {
        "tags": [
          "Chat Integration"
        ],
        "summary": "POST /api/v1/integrations/chat/events",
        "description": "Webhook receiver for chat events from external chat systems.\nVerifies the event signature and processes the event.",
        "operationId": "receive_chat_event",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReceiveChatEventRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Event processed"
          },
          "401": {
            "description": "Missing or invalid signature",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/integrations/chat/unfurl": {
      "post": {
        "tags": [
          "Chat Integration"
        ],
        "summary": "POST /api/v1/integrations/chat/unfurl",
        "description": "Link unfurl endpoint that accepts a RustShare URL and returns preview metadata.\nRequires authentication to verify the user has permission to view the shared resource.",
        "operationId": "unfurl_link",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnfurlLinkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Unfurled link metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnfurlLinkResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Permission denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Share or file not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/integrations/chat/unfurl/public": {
      "post": {
        "tags": [
          "Chat Integration"
        ],
        "summary": "POST /api/v1/integrations/chat/unfurl/public",
        "description": "Public link unfurl endpoint that doesn't require authentication.\nThis is for chat systems that want to preview public shares.",
        "operationId": "unfurl_link_public",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnfurlLinkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Unfurled link metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnfurlLinkResponse"
                }
              }
            }
          },
          "404": {
            "description": "Share or file not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/integrations/webhooks/dispatch": {
      "post": {
        "tags": [
          "Chat Integration"
        ],
        "summary": "POST /api/v1/integrations/webhooks/dispatch",
        "description": "Internal endpoint to dispatch events to registered chat webhooks.\nThis is called internally when shares are revoked or files change.",
        "operationId": "dispatch_webhooks",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DispatchWebhookRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Dispatch results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DispatchWebhookResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DispatchWebhookRequest": {
        "type": "object",
        "description": "Request to dispatch an event to registered chat webhooks.",
        "required": [
          "event_type",
          "payload"
        ],
        "properties": {
          "event_type": {
            "type": "string"
          },
          "payload": {}
        }
      },
      "DispatchWebhookResponse": {
        "type": "object",
        "description": "Response from dispatching webhooks.",
        "required": [
          "dispatched"
        ],
        "properties": {
          "dispatched": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WebhookDispatchResult"
            }
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Standard error response format.",
        "required": [
          "error"
        ],
        "properties": {
          "details": {
            "type": [
              "string",
              "null"
            ]
          },
          "error": {
            "type": "string"
          }
        }
      },
...
      "IncomingChatEvent": {
        "type": "object",
        "description": "Incoming chat event from external chat system.",
        "required": [
          "event_type",
          "timestamp",
          "data"
        ],
        "properties": {
          "data": {},
          "event_type": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "user_id": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "ReceiveChatEventRequest": {
        "type": "object",
        "description": "Request to receive a chat event from an external system.",
        "required": [
          "event"
        ],
        "properties": {
          "event": {
            "$ref": "#/components/schemas/IncomingChatEvent"
          }
        }
      },
...
      "RegisterWebhookRequest": {
        "type": "object",
        "description": "Request to register a new webhook URL.",
        "required": [
          "url"
        ],
        "properties": {
          "url": {
            "type": "string"
          }
        }
      },
...
      "UnfurlLinkRequest": {
        "type": "object",
        "description": "Request to unfurl a RustShare link.",
        "required": [
          "url"
        ],
        "properties": {
          "url": {
            "type": "string"
          }
        }
      },
      "UnfurlLinkResponse": {
        "type": "object",
        "description": "Response from unfurling a link.",
        "required": [
          "title",
          "resource_type",
          "resource_id",
          "share_token",
          "created_at",
          "password_protected",
          "permissions"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "expires_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "mime_type": {
            "type": [
              "string",
              "null"
            ]
          },
          "password_protected": {
            "type": "boolean"
          },
          "permissions": {
            "type": "string"
          },
          "resource_id": {
            "type": "string",
            "format": "uuid"
          },
          "resource_type": {
            "type": "string"
          },
          "share_token": {
            "type": "string"
          },
          "size": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "thumbnail_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "title": {
            "type": "string"
          }
        }
      },
...
      "WebhookDispatchResult": {
        "type": "object",
        "description": "Result of a single webhook dispatch.",
        "required": [
          "url",
          "success"
        ],
        "properties": {
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "success": {
            "type": "boolean"
          },
          "url": {
            "type": "string"
          }
        }
      },
      "WebhookListResponse": {
        "type": "object",
        "description": "Response containing registered webhooks.",
        "required": [
          "webhooks"
        ],
        "properties": {
          "webhooks": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  },
...
}